Fixes bug opening odd length GBKs with a space at end
In the edge case of a Genbank file with an odd length sequence and a
blank line after the sequence, preceeding the final set of double
slashes, the commented out function was called. However, the array "dna"
passed to it was empty, triggering an ArrayIndexOutOfBoundsException. I
could see no other cases when this if was triggered (in even and odd
length sequences, with and without a newline before the end), and so
commented it out. Fixes that issue.